--- a/src/dev/com/yetanalytics/flint/sparql.clj
+++ b/src/dev/com/yetanalytics/flint/sparql.clj
@@ -1,5 +1,6 @@
 (ns com.yetanalytics.flint.sparql
-  (:require [clojure.java.io :as io])
+  (:require [clojure.java.io :as io]
+            [criterium.core :as crit])
   (:import [org.apache.jena.query QueryFactory]
            [org.apache.jena.update UpdateFactory]))
 
@@ -155,8 +156,7 @@
     DROP SILENT ALL;"))
 
 (comment
-  (require '[com.yetanalytics.flint :as flint]
-           '[criterium.core :as crit])
+  (require '[com.yetanalytics.flint :as flint])
 
   (crit/quick-bench
    (flint/format-query
@@ -186,3 +186,44 @@
              ['?x :foo/ttttttttten "TEN"]]}
     :validate? false))
   )
+
+(comment
+  (require '[com.yetanalytics.flint.axiom.impl.vaidation :as v])
+
+  ;; ASCII bench
+  (crit/quick-bench
+   (re-matches #"\?\w+" (name '?supercalifragilisticexpialidocious)))
+
+  (crit/quick-bench
+   (re-matches v/var-regex (name '?supercalifragilisticexpialidocious)))
+
+  (crit/bench
+   (re-matches v/var-regex (name '?x)))
+
+  (crit/quick-bench
+   (v/valid-var-symbol? '?supercalifragilisticexpialidocious))
+
+  (crit/bench
+   (v/valid-var-symbol? '?x))
+
+  ;; Non-ASCII bench
+  (crit/quick-bench
+   (re-matches v/var-regex (name '?我们去西安吃面)))
+
+  (crit/quick-bench
+   (v/valid-var-symbol? '?我们今天吃西安炒面))
+
+  ;; String literal bench
+  (crit/quick-bench
+   (re-matches v/literal-regex
+               "\\\"supercalifragilisticexpialidocious\\\""))
+
+  (crit/quick-bench
+   (v/valid-literal-str? "\\\"supercalifragilisticexpialidocious\\\""))
+
+  ;; Prefix name bench
+  (crit/bench
+   (re-matches v/prefix-name-regex (name :foo%80bar)))
+
+  (crit/bench
+   (v/valid-prefix-keyword? :foo%80bar)))