--- a/src/main/com/yetanalytics/flint/axiom/impl/validation.cljc
+++ b/src/main/com/yetanalytics/flint/axiom/impl/validation.cljc
@@ -1,7 +1,86 @@
-(ns com.yetanalytics.flint.axiom.impl.validation)
+(ns com.yetanalytics.flint.axiom.impl.validation
+  #?(:cljs (:require [goog.string :refer [format]]
+                     [goog.string.format])))
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
-;; Regex
+;; Ranges
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;; 
+
+(defn- char->int
+  [c]
+  #?(:clj (int c) :cljs (.charCodeAt c)))
+
+(def qmark-range [(char->int \?)])
+(def uscore-range [(char->int \_)])
+
+(def unicode-range-no-digits
+  [[(char->int \A) (char->int \Z)]
+   [(char->int \a) (char->int \z)]
+   [0x0060 0x007B] ; * Latin letters w/ diacritics (excl. multiplication and
+   [0x00C0 0x00D6] ;   division signs) + spacing modifiers
+   [0x00D8 0x00F6] ;
+   [0x0370 0x037D] ; * Various alphabets (excl. Greek question mark)
+   [0x037F 0x1FFF] ;
+   [0x200C 0x200D] ; * Zero-width Joiner + Non-joiner
+   [0x2070 0x218F] ; * Super/sub-scripts + other punctuation
+   [0x2C00 0x2FEF] ; * More alphabets + Punctuation + Hanzi radicals
+   [0x3001 0xD7FF] ; * CJK punctuation + hanzi + hangeul
+   [0xF900 0xFDCF] ; * More CJK characters + Hebrew + Arabic
+   [0xFDF0 0xFFFD]])
+
+;; The Unicode code points \u10000-\uEFFFF are permitted in the SPARQL spec,
+;; but are unsupported here since Java only supports chars up to \uFFFF.
+;; In addition, the spec prohibits surrogate pair chars (\uD800-\uF8FF) so
+;; only Basic Multilingual Plane chars are allowed.
+;; This is okay since it's doubtful people will be using emojis and other
+;; supplemental plane chars in Flint.
+
+(def unicode-range-start
+  (into [[(char->int \0) (char->int \9)] (char->int \_)]
+        unicode-range-no-digits))
+
+;; \u00B7        = middle dot ·
+;; \u0300-\u036F = combining characters
+;; \u203F-\u2040 = ties
+(def unicode-range
+  (into [0x00B7 [0x0300 0x036F] [0x203F 0x2040]]
+        unicode-range-start))
+
+(def var-start-range
+  unicode-range-start)
+
+(def var-body-range
+  unicode-range)
+
+(def bnode-start-range
+  unicode-range-start)
+
+(def bnode-body-range
+  (into [(char->int \-) (char->int \.)] unicode-range))
+
+(def bnode-end-range
+  (into [(char->int \-)] unicode-range))
+
+(def prefix-ns-start-range
+  unicode-range-no-digits)
+
+(def prefix-ns-body-range
+  (into [(char->int \-) (char->int \.)] unicode-range))
+
+(def prefix-ns-end-range
+  (into [(char->int \-)] unicode-range))
+
+(def prefix-name-start-range
+  (into [(char->int \:)] unicode-range-start))
+
+(def prefix-name-body-range
+  (into [(char->int \:) (char->int \-) (char->int \.)] unicode-range))
+
+(def prefix-name-end-range
+  (into [(char->int \:) (char->int \-)] unicode-range))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Regexes
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;; 
 
 (def iri-regex*
@@ -10,28 +89,248 @@
 (def iri-regex
   #"<([^<>{}\"\\|\^`\s])*>")
 
-(def prefix-iri-ns-regex
-  #"[A-Za-z_]([\w\-\.]*[\w\-])?")
-
-(def prefix-iri-name-regex
-  #"[\w\-]+")
-
-(def variable-regex
-  #"\?\w+")
-
-(def bnode-regex
-  #"_(\w([\w\.]*\w)?)?")
-
 ;; Note in the second part that we need to match the letters `n`, `r`, etc.,
 ;; not the newline char `\n` nor the return char `\r`.
 (def valid-str-regex
   #"([^\"\r\n\\]|(?:\\(?:t|b|n|r|f|\\|\"|')))*")
+
+(defn- num->unicode-str
+  [n]
+  #?(:clj
+     (cond
+       (<= n 0x000F)
+       (format "\\u000%x" n)
+       (<= n 0x00FF)
+       (format "\\u00%x" n)
+       (<= n 0x0FFF)
+       (format "\\u0%x" n)
+       :else
+       (format "\\u%x" n))
+     :cljs
+     (let [hex-str (.toUpperCase (.toString ^number n 16))]
+       (cond
+         (<= n 0x000F)
+         (format "\\u000%s" hex-str)
+         (<= n 0x00FF)
+         (format "\\u00%s" hex-str)
+         (<= n 0x0FFF)
+         (format "\\u0%s" hex-str)
+         :else
+         (format "\\u%s" hex-str)))))
+
+(defn- ranges->regex-charset
+  [ranges]
+  (let [charset
+        (reduce (fn [s r]
+                  (if (vector? r)
+                    (let [[start end] r
+                          start-str (num->unicode-str start)
+                          end-str   (num->unicode-str end)]
+                      (str s start-str "-" end-str))
+                    (let [range-str (num->unicode-str r)]
+                      (str s range-str))))
+                ""
+                ranges)]
+    (str "[" charset "]")))
+
+(def var-regex
+  (let [var-start (ranges->regex-charset var-start-range)
+        var-body  (ranges->regex-charset var-body-range)]
+    (re-pattern (format "\\?%s%s*"
+                        var-start
+                        var-body))))
+
+(def bnode-regex
+  (let [bnode-start (ranges->regex-charset bnode-start-range)
+        bnode-body  (ranges->regex-charset bnode-body-range)
+        bnode-end   (ranges->regex-charset bnode-end-range)]
+    (re-pattern (format "\\_(?:%s(?:%s*%s)?)?"
+                        bnode-start
+                        bnode-body
+                        bnode-end))))
+
+(def prefix-ns-regex
+  (let [prefix-ns-start (ranges->regex-charset prefix-ns-start-range)
+        prefix-ns-body  (ranges->regex-charset prefix-ns-body-range)
+        prefix-ns-end   (ranges->regex-charset prefix-ns-end-range)]
+    (re-pattern (format "%s(?:%s*%s)?"
+                        prefix-ns-start
+                        prefix-ns-body
+                        prefix-ns-end))))
+
+(def prefix-name-regex
+  (let [prefix-name-start (ranges->regex-charset prefix-name-start-range)
+        prefix-name-body  (ranges->regex-charset prefix-name-body-range)
+        prefix-name-end   (ranges->regex-charset prefix-name-end-range)]
+    (re-pattern (format "%s(?:%s*%s)?"
+                        prefix-name-start
+                        prefix-name-body
+                        prefix-name-end))))
+
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
+;; Clojure-specific validation
+;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;; 
+
+#?(:clj
+   (defmacro unicode-bitset
+     "Instantiate a `java.util.BitSet` instance and add set all bits
+      at or (inclusive) between the ranges in `code-point-ranges`."
+     [code-point-ranges]
+     (let [cp-ranges# (if (symbol? code-point-ranges)
+                        @(ns-resolve *ns* code-point-ranges)
+                        code-point-ranges)]
+       `(doto (java.util.BitSet. 0xFFFF)
+          ~@(map (fn [r#]
+                   (if (vector? r#)
+                     `(.set ~(first r#) ~(inc (second r#)))
+                     `(.set ~r#)))
+                 cp-ranges#)))))
+
+#?(:clj (def ^java.util.BitSet qmark-bitset
+          (unicode-bitset qmark-range)))
+#?(:clj (def ^java.util.BitSet uscore-bitset
+          (unicode-bitset uscore-range)))
+
+#?(:clj (def ^java.util.BitSet var-start-bitset
+          (unicode-bitset var-start-range)))
+#?(:clj (def ^java.util.BitSet var-body-bitset
+          (unicode-bitset var-body-range)))
+
+#?(:clj (def ^java.util.BitSet bnode-start-bitset
+          (unicode-bitset bnode-start-range)))
+#?(:clj (def ^java.util.BitSet bnode-body-bitset
+          (unicode-bitset bnode-body-range)))
+#?(:clj (def ^java.util.BitSet bnode-end-bitset
+          (unicode-bitset bnode-end-range)))
+
+#?(:clj (def ^java.util.BitSet prefix-ns-start-bitset
+          (unicode-bitset prefix-ns-start-range)))
+#?(:clj (def ^java.util.BitSet prefix-ns-body-bitset
+          (unicode-bitset prefix-ns-body-range)))
+#?(:clj (def ^java.util.BitSet prefix-ns-end-bitset
+          (unicode-bitset prefix-ns-end-range)))
+
+#?(:clj (def ^java.util.BitSet prefix-name-start-bitset
+          (unicode-bitset prefix-name-start-range)))
+#?(:clj (def ^java.util.BitSet prefix-name-body-bitset
+          (unicode-bitset prefix-name-body-range)))
+#?(:clj (def ^java.util.BitSet prefix-name-end-bitset
+          (unicode-bitset prefix-name-end-range)))
+
+#?(:clj
+   (defmacro in-bitset?
+     "Inlined check on whether the code point at `idx` in `s` can be found
+      in `bitset`."
+     [bitset s idx]
+     `(.get ^java.util.BitSet ~bitset (.codePointAt ^String ~s ~idx))))
+
+#?(:clj
+   (defmacro recur-if
+     "Perform a recursion only when `pred` evaluates to `true`; otherwise,
+      return `false` and do not recur."
+     [pred & exprs]
+     `(if ~pred
+        (recur ~@exprs)
+        false)))
+
+#?(:clj
+   #_{:clj-kondo/ignore [:loop-without-recur]}
+   (defn- valid-var-str?*
+     [^String var-str]
+     (let [ccnt (.codePointCount var-str 0 (count var-str))]
+       (loop [idx 0]
+         (cond
+           (>= idx ccnt)
+           (<= 2 ccnt) ; Need to have initial qmark + at least one start char
+           (= idx 0)
+           (recur-if (in-bitset? qmark-bitset var-str idx)
+                     (inc idx))
+           (= idx 1)
+           (recur-if (in-bitset? var-start-bitset var-str idx)
+                     (inc idx))
+           :else
+           (recur-if (in-bitset? var-body-bitset var-str idx)
+                     (inc idx)))))))
+
+#?(:clj
+   #_{:clj-kondo/ignore [:loop-without-recur]}
+   (defn- valid-bnode-str?*
+     [^String bnode-str]
+     (let [ccnt (.codePointCount bnode-str 0 (count bnode-str))
+           lidx (dec ccnt)]
+       (loop [idx 0]
+         (cond
+           (>= idx ccnt)
+           (<= 1 ccnt) ; Need to have at least the initial underscore
+           (= idx 0)
+           (recur-if (in-bitset? uscore-bitset bnode-str idx)
+                     (inc idx))
+           (= idx 1)
+           (recur-if (in-bitset? bnode-start-bitset bnode-str idx)
+                     (inc idx))
+           (= idx lidx)
+           (recur-if (in-bitset? bnode-end-bitset bnode-str idx)
+                     (inc idx))
+           :else
+           (recur-if (in-bitset? bnode-body-bitset bnode-str idx)
+                     (inc idx)))))))
+
+#?(:clj
+   #_{:clj-kondo/ignore [:loop-without-recur]}
+   (defn- valid-prefix-ns-str?*
+     [^String prefix-ns-str]
+     (let [ccnt (.codePointCount prefix-ns-str 0 (count prefix-ns-str))
+           lidx (dec ccnt)]
+       (loop [idx 0]
+         (cond
+           (>= idx ccnt)
+           (<= 1 ccnt) ; Need to have at least one char
+           (= idx 0)
+           (recur-if (in-bitset? prefix-ns-start-bitset prefix-ns-str idx)
+                     (inc idx))
+           (= idx lidx)
+           (recur-if (in-bitset? prefix-ns-end-bitset prefix-ns-str idx)
+                     (inc idx))
+           :else
+           (recur-if (in-bitset? prefix-ns-body-bitset prefix-ns-str idx)
+                     (inc idx)))))))
+
+
+#?(:clj
+   #_{:clj-kondo/ignore [:loop-without-recur]}
+   (defn- valid-prefix-name-str?*
+     [^String prefix-name-str]
+     (let [ccnt (.codePointCount prefix-name-str 0 (count prefix-name-str))
+           lidx (dec ccnt)]
+       (loop [idx 0]
+         (cond
+           (>= idx ccnt)
+           (<= 1 ccnt) ; Need to have at least one char
+           (= idx 0)
+           (recur-if (in-bitset? prefix-name-start-bitset prefix-name-str idx)
+                     (inc idx))
+           (= idx lidx)
+           (recur-if (in-bitset? prefix-name-end-bitset prefix-name-str idx)
+                     (inc idx))
+           :else
+           (recur-if (in-bitset? prefix-name-body-bitset prefix-name-str idx)
+                     (inc idx)))))))
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; Validation
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 
 ;; IRIs and RDF terms
+
+(defn- valid-prefix-ns-str?
+  [prefix-ns-str]
+  #?(:clj (valid-prefix-ns-str?* prefix-ns-str)
+     :cljs (boolean (re-matches prefix-ns-regex prefix-ns-str))))
+
+(defn- valid-prefix-name-str?
+  [prefix-name-str]
+  #?(:clj (valid-prefix-name-str?* prefix-name-str)
+     :cljs (boolean (re-matches prefix-name-regex prefix-name-str))))
 
 (defn valid-iri-string?*
   "Is `x` an IRI string? Note that `x` can be an otherwise non-IRI
@@ -49,7 +348,7 @@
   "Is `k` a valid SPARQL prefix keyword?"
   [k]
   (boolean (and (nil? (namespace k))
-                (or (re-matches prefix-iri-ns-regex (name k))
+                (or (valid-prefix-name-str? (name k))
                     (= :$ k)))))
 
 (defn valid-prefix-iri-keyword?
@@ -58,20 +357,29 @@
   (let [kns   (namespace k)
         kname (name k)]
     (boolean (and (not (#{:a :*} k))
-                  (or (nil? kns)
-                      (re-matches prefix-iri-ns-regex kns))
-                  (re-matches prefix-iri-name-regex kname)))))
+                  (or (nil? kns) (valid-prefix-ns-str? kns))
+                  (valid-prefix-name-str? kname)))))
+
+(defn- valid-var-str?
+  [var-str]
+  #?(:clj (valid-var-str?* var-str)
+     :cljs (boolean (re-matches var-regex var-str))))
+
+(defn- valid-bnode-str?
+  [bnode-str]
+  #?(:clj (valid-bnode-str?* bnode-str)
+     :cljs (boolean (re-matches bnode-regex bnode-str))))
 
 (defn valid-var-symbol?
   "Is `var-sym` a symbol that starts with `?`?"
   [var-sym]
-  (boolean (re-matches variable-regex (name var-sym))))
+  (valid-var-str? (name var-sym)))
 
 (defn valid-bnode-symbol?
   "Is `bnode-sym` a symbol that starts with `_` and has zero or more
    trailing chars?"
   [bnode-sym]
-  (boolean (re-matches bnode-regex (name bnode-sym))))
+  (valid-bnode-str? (name bnode-sym)))
 
 ;; Literals
 
@@ -87,3 +395,23 @@
   (boolean (and (->> lang-map count (= 1))
                 (->> lang-map keys first keyword?)
                 (->> lang-map vals first (re-matches valid-str-regex)))))
+
+(comment
+  (require '[criterium.core :as crit])
+
+  ;; ASCII bench
+  (crit/quick-bench
+   (re-matches #"\?\w+" (name '?supercalifragilisticexpialidocious)))
+
+  (crit/quick-bench
+   (re-matches var-regex (name '?supercalifragilisticexpialidocious)))
+  
+  (crit/quick-bench
+   (valid-var-symbol? '?supercalifragilisticexpialidocious))
+
+  ;; Non-ASCII bench
+  (crit/quick-bench
+   (re-matches var-regex (name '?我们去西安吃面)))
+
+  (crit/quick-bench
+   (valid-var-symbol? '?我们今天吃西安炒面)))