… doc/axiom.md