- Add support for Unicode (Basic Multilingual Plane) characters in variables, blank nodes, and prefixed IRIs.
- Add support for percent encoding for prefixed IRIs.
- Ensure that language tags are alphanumeric strings, as per the spec.
- Optimize validation (in Clojure) using low-level Java constructs.

Benching the results of Clojure optimization:
```clojure
(require '[com.yetanalytics.flint.axiom.impl.validation :as v])

;; Before: Regex WITHOUT Unicode
(crit/quick-bench
 (re-matches #"\?\w+" (name '?foo)))
;; =>  98.5 ns

;; During: Regex WITH Unicode
(crit/quick-bench
 (re-matches v/var-regex (name '?foo)))
;; => 276.9 ns

;; After: New function WITH Unicode
(crit/quick-bench
 (v/valid-var-symbol? '?foo))
;; =>  67.3 ns
```
On longer strings the difference between regex and custom Unicode validation become more apparent:
```clojure
(def v '?supercalifragilisticexpialidocious)

(crit/quick-bench
 (re-matches #"\?\w+" (name v)))
;; => 335.6 ns

(crit/quick-bench
 (re-matches v/var-regex (name v)))
;; => 2522.1 ns

(crit/quick-bench
 (v/valid-var-symbol? v))
;; => 348.4 ns
```